requirements.txt
ultralytics==8.3.79
opencv-python==4.7.0.72
Flask==3.1.0
Pillow==9.4.0
Werkzeug==3.1.3